+++ src/config.js
@@ -1,109 +1,37 @@
 const core = require('@actions/core')
 const github = require('@actions/github')
-const parser = require('action-input-parser')
-require('dotenv').config()
 
 const IS_PR = [ 'pull_request', 'pull_request_target' ].includes(github.context.eventName)
 
 const context = {
-	GITHUB_TOKEN: parser.getInput({
-		key: [ 'GH_PAT', 'GITHUB_TOKEN' ],
-		required: true
-	}),
-	VERCEL_TOKEN: parser.getInput({
-		key: 'VERCEL_TOKEN',
-		required: true
-	}),
-	VERCEL_ORG_ID: parser.getInput({
-		key: 'VERCEL_ORG_ID',
-		required: true
-	}),
-	VERCEL_PROJECT_ID: parser.getInput({
-		key: 'VERCEL_PROJECT_ID',
-		required: true
-	}),
-	PRODUCTION: parser.getInput({
-		key: 'PRODUCTION',
-		type: 'boolean',
-		default: !IS_PR
-	}),
-	GITHUB_DEPLOYMENT: parser.getInput({
-		key: 'GITHUB_DEPLOYMENT',
-		type: 'boolean',
-		default: true
-	}),
-	CREATE_COMMENT: parser.getInput({
-		key: 'CREATE_COMMENT',
-		type: 'boolean',
-		default: true
-	}),
-	DELETE_EXISTING_COMMENT: parser.getInput({
-		key: 'DELETE_EXISTING_COMMENT',
-		type: 'boolean',
-		default: true
-	}),
-	ATTACH_COMMIT_METADATA: parser.getInput({
-		key: 'ATTACH_COMMIT_METADATA',
-		type: 'boolean',
-		default: true
-	}),
-	DEPLOY_PR_FROM_FORK: parser.getInput({
-		key: 'DEPLOY_PR_FROM_FORK',
-		type: 'boolean',
-		default: false
-	}),
-	PR_LABELS: parser.getInput({
-		key: 'PR_LABELS',
-		default: [ 'deployed' ],
-		type: 'array',
-		disableable: true
-	}),
-	ALIAS_DOMAINS: parser.getInput({
-		key: 'ALIAS_DOMAINS',
-		type: 'array',
-		disableable: true
-	}),
-	PR_PREVIEW_DOMAIN: parser.getInput({
-		key: 'PR_PREVIEW_DOMAIN'
-	}),
-	VERCEL_SCOPE: parser.getInput({
-		key: 'VERCEL_SCOPE'
-	}),
-	GITHUB_REPOSITORY: parser.getInput({
-		key: 'GITHUB_REPOSITORY',
-		required: true
-	}),
-	GITHUB_DEPLOYMENT_ENV: parser.getInput({
-		key: 'GITHUB_DEPLOYMENT_ENV'
-	}),
-	TRIM_COMMIT_MESSAGE: parser.getInput({
-		key: 'TRIM_COMMIT_MESSAGE',
-		type: 'boolean',
-		default: false
-	}),
-	WORKING_DIRECTORY: parser.getInput({
-		key: 'WORKING_DIRECTORY'
-	}),
-	BUILD_ENV: parser.getInput({
-		key: 'BUILD_ENV',
-		type: 'array'
-	}),
-	PREBUILT: parser.getInput({
-		key: 'PREBUILT',
-		type: 'boolean',
-		default: false
-	}),
+	GITHUB_TOKEN: core.getInput('GITHUB_TOKEN', { required: true }),
+	VERCEL_TOKEN: core.getInput('VERCEL_TOKEN', { required: true }),
+	VERCEL_ORG_ID: core.getInput('VERCEL_ORG_ID', { required: true }),
+	VERCEL_PROJECT_ID: core.getInput('VERCEL_PROJECT_ID', { required: true }),
+	PRODUCTION: core.getBooleanInput('PRODUCTION', { required: false }),
+	GITHUB_DEPLOYMENT: core.getBooleanInput('GITHUB_DEPLOYMENT', { required: false }),
+	CREATE_COMMENT: core.getBooleanInput('CREATE_COMMENT', { required: false }),
+	DELETE_EXISTING_COMMENT: core.getBooleanInput('DELETE_EXISTING_COMMENT', { required: false }),
+	ATTACH_COMMIT_METADATA: core.getBooleanInput('ATTACH_COMMIT_METADATA', { required: false }),
+	DEPLOY_PR_FROM_FORK: core.getBooleanInput('DEPLOY_PR_FROM_FORK', { required: false }),
+	PR_LABELS: core.getMultilineInput('PR_LABELS', { required: false }),
+	ALIAS_DOMAINS: core.getMultilineInput('ALIAS_DOMAINS', { required: false }),
+	PR_PREVIEW_DOMAIN: core.getInput('PR_PREVIEW_DOMAIN'),
+	VERCEL_SCOPE: core.getInput('VERCEL_SCOPE'),
+	GITHUB_DEPLOYMENT_ENV: core.getInput('GITHUB_DEPLOYMENT_ENV'),
+	TRIM_COMMIT_MESSAGE: core.getBooleanInput('TRIM_COMMIT_MESSAGE', { required: false }),
+	WORKING_DIRECTORY: core.getInput('WORKING_DIRECTORY'),
+	BUILD_ENV: core.getMultilineInput('BUILD_ENV', { required: false }),
+	RUNTIME_ENV: core.getMultilineInput('RUNTIME_ENV', { required: false }),
+	PREBUILT: core.getBooleanInput('PREBUILT', { required: false }),
 	RUNNING_LOCAL: process.env.RUNNING_LOCAL === 'true',
-	FORCE: parser.getInput({
-		key: 'FORCE',
-		type: 'boolean',
-		default: false
-	})
+	FORCE: core.getBooleanInput('FORCE', { required: false })
 }
 
 const setDynamicVars = () => {
-	context.USER = context.GITHUB_REPOSITORY.split('/')[0]
-	context.REPOSITORY = context.GITHUB_REPOSITORY.split('/')[1]
+	const { owner, repo } = github.context.repo
+	context.USER = owner
+	context.REPOSITORY = repo
 
 	// If running the action locally, use env vars instead of github.context
 	if (context.RUNNING_LOCAL) {
@@ -112,7 +40,6 @@ const setDynamicVars = () => {
 		context.PR_NUMBER = process.env.PR_NUMBER || undefined
 		context.REF = process.env.REF || 'refs/heads/master'
 		context.BRANCH = process.env.BRANCH || 'master'
-		context.PRODUCTION = process.env.PRODUCTION === 'true' || !context.IS_PR
 		context.LOG_URL = process.env.LOG_URL || `https://github.com/${ context.USER }/${ context.REPOSITORY }`
 		context.ACTOR = process.env.ACTOR || context.USER
 		context.IS_FORK = process.env.IS_FORK === 'true' || false
@@ -131,12 +58,12 @@ const setDynamicVars = () => {
 		context.REF = github.context.payload.pull_request.head.ref
 		context.SHA = github.context.payload.pull_request.head.sha
 		context.BRANCH = github.context.payload.pull_request.head.ref
-		context.IS_FORK = github.context.payload.pull_request.head.repo.full_name !== context.GITHUB_REPOSITORY
+		context.IS_FORK = github.context.payload.pull_request.head.repo.full_name !== process.env.GITHUB_REPOSITORY
 	} else {
 		context.ACTOR = github.context.actor
 		context.REF = github.context.ref
 		context.SHA = github.context.sha
-		context.BRANCH = github.context.ref.substr(11)
+		context.BRANCH = github.context.ref.substring('refs/heads/'.length)
 	}
 }
 
@@ -146,11 +73,7 @@ core.setSecret(context.GITHUB_TOKEN)
 core.setSecret(context.VERCEL_TOKEN)
 
 core.debug(
-	JSON.stringify(
-		context,
-		null,
-		2
-	)
+	JSON.stringify(context)
 )
 
 module.exports = context